+++ kmath-geometry/src/commonMain/kotlin/space/kscience/kmath/geometry/Line.kt
@@ -12,16 +12,16 @@ import kotlinx.serialization.Serializable
  * A line formed by [start] vector of start and a [direction] vector. Direction vector is not necessarily normalized,
  * but its length does not affect line properties
  */
-public interface Line<out V : Vector> {
+public interface Line<out V : Any> {
     public val start: V
     public val direction: V
 }
 
 @Serializable
 @SerialName("Line")
-private data class LineImpl<out V : Vector>(override val start: V, override val direction: V): Line<V>
+private data class LineImpl<out V : Any>(override val start: V, override val direction: V) : Line<V>
 
-public fun <V : Vector> Line(base: V, direction: V): Line<V> = LineImpl(base, direction)
+public fun <V : Any> Line(base: V, direction: V): Line<V> = LineImpl(base, direction)
 
 public typealias Line2D = Line<DoubleVector2D>
 public typealias Line3D = Line<DoubleVector3D>
@@ -29,7 +29,7 @@ public typealias Line3D = Line<DoubleVector3D>
 /**
  * A directed line segment between [begin] and [end]
  */
-public interface LineSegment<out V : Vector> {
+public interface LineSegment<out V : Any> {
     public val begin: V
     public val end: V
 }
@@ -39,11 +39,11 @@ public interface LineSegment<out V : Vector> {
  */
 @Serializable
 @SerialName("LineSegment")
-private data class LineSegmentImpl<out V : Vector>(override val begin: V, override val end: V) : LineSegment<V>
+private data class LineSegmentImpl<out V : Any>(override val begin: V, override val end: V) : LineSegment<V>
 
-public fun <V : Vector> LineSegment(begin: V, end: V): LineSegment<V> = LineSegmentImpl(begin, end)
+public fun <V : Any> LineSegment(begin: V, end: V): LineSegment<V> = LineSegmentImpl(begin, end)
 
-public fun <V : Vector> LineSegment<V>.line(algebra: GeometrySpace<V>): Line<V> = with(algebra) {
+public fun <V : Any> LineSegment<V>.line(algebra: GeometrySpace<V, *>): Line<V> = with(algebra) {
     Line(begin, end - begin)
 }