+++ kmath-geometry/src/commonMain/kotlin/space/kscience/kmath/geometry/projections.kt
@@ -12,7 +12,7 @@ package space.kscience.kmath.geometry
  * @param vector to project
  * @param line line to which vector should be projected
  */
-public fun <V : Vector> GeometrySpace<V>.projectToLine(vector: V, line: Line<V>): V = with(line) {
+public fun <V : Any> GeometrySpace<V, *>.projectToLine(vector: V, line: Line<V>): V = with(line) {
     start + (direction dot (vector - start)) / (direction dot direction) * direction
 }
 
@@ -23,5 +23,5 @@ public fun <V : Vector> GeometrySpace<V>.projectToLine(vector: V, line: Line<V>)
  * @param normal normal (perpendicular) vector to a hyper-plane to which vector should be projected
  * @param base point belonging to a hyper-plane to which vector should be projected
  */
-public fun <V : Vector> GeometrySpace<V>.projectAlong(vector: V, normal: V, base: V): V =
+public fun <V : Any> GeometrySpace<V, *>.projectAlong(vector: V, normal: V, base: V): V =
     vector + normal * ((base - vector) dot normal) / (normal dot normal)