+++ kmath-geometry/src/jvmMain/kotlin/space/kscience/kmath/geometry/lineExtensions.kt
@@ -6,15 +6,16 @@
 import space.kscience.kmath.geometry.GeometrySpace
 import space.kscience.kmath.geometry.Line
 import space.kscience.kmath.geometry.LineSegment
-import space.kscience.kmath.geometry.Vector
 import space.kscience.kmath.operations.Group
 
 /**
  * Get a line, containing this [LineSegment]
  */
-context(Group<V>) public val <V : Vector> LineSegment<V>.line: Line<V> get() = Line(begin, end - begin)
+context(Group<V>)
+public val <V : Any> LineSegment<V>.line: Line<V> get() = Line(begin, end - begin)
 
 /**
  * Get a length of a line segment
  */
-context(GeometrySpace<V>) public val <V : Vector> LineSegment<V>.length: Double get() = norm(end - begin)
+context(GeometrySpace<V>)
+public val <V : Any> LineSegment<V>.length: Double get() = norm(end - begin)